[#1748] neofs-adm: Move `dump-hashes` command to a separate file
I have just lost 3 minutes trying to find it.
